requirements.txt
mysql-connector
azure-functions
numpy
joblib
pandas